refactor: rename dao packages to streamline structure
