+++ src/main/java/tech/kp45/bids/bridge/AppStartedEvent.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Component;
 
 import lombok.extern.slf4j.Slf4j;
 import tech.kp45.bids.bridge.dataset.Dataset;
-import tech.kp45.bids.bridge.dataset.service.DatasetService;
+import tech.kp45.bids.bridge.dataset.DatasetService;
 import tech.kp45.bids.bridge.job.Job;
 import tech.kp45.bids.bridge.job.service.JobService;
 import tech.kp45.bids.bridge.pipeline.Pipeline;
