+++ src/main/java/tech/kp45/bids/bridge/job/JobMapper.java
@@ -1,8 +1,6 @@
-package tech.kp45.bids.bridge.job.dao;
+package tech.kp45.bids.bridge.job;
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-
-import tech.kp45.bids.bridge.job.Job;
 
 public interface JobMapper extends BaseMapper<Job> {
