+++ src/main/java/tech/kp45/bids/bridge/job/service/JobService.java
@@ -10,10 +10,10 @@ import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import lombok.extern.slf4j.Slf4j;
 import tech.kp45.bids.bridge.common.exception.BasicRuntimeException;
 import tech.kp45.bids.bridge.dataset.Dataset;
-import tech.kp45.bids.bridge.dataset.service.DatasetService;
+import tech.kp45.bids.bridge.dataset.DatasetService;
 import tech.kp45.bids.bridge.job.Job;
+import tech.kp45.bids.bridge.job.JobMapper;
 import tech.kp45.bids.bridge.job.JobStatus;
-import tech.kp45.bids.bridge.job.dao.JobMapper;
 import tech.kp45.bids.bridge.job.scheduler.JobEngine;
 import tech.kp45.bids.bridge.pipeline.Pipeline;
 import tech.kp45.bids.bridge.pipeline.PipelineService;
