+++ src/main/java/tech/kp45/bids/bridge/pipeline/PipelineMapper.java
@@ -1,8 +1,6 @@
-package tech.kp45.bids.bridge.pipeline.dao;
+package tech.kp45.bids.bridge.pipeline;
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-
-import tech.kp45.bids.bridge.pipeline.Pipeline;
 
 public interface PipelineMapper extends BaseMapper<Pipeline> {
