+++ src/main/java/tech/kp45/bids/bridge/pipeline/PipelineService.java
@@ -5,8 +5,6 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import tech.kp45.bids.bridge.pipeline.dao.PipelineMapper;
-
 @Service
 public class PipelineService {
